requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
email-validator==2.1.0.post1
textblob==0.17.1
numpy==1.26.3
alembic==1.13.1
Pillow==10.2.0
requests==2.31.0
matplotlib==3.8.2
pandas==2.1.3
scikit-learn==1.3.2
python-dateutil==2.8.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
blinker==1.7.0
reportlab==4.4.1 